requirements.txt
Django==1.9.1
psycopg2==2.6.1
django-celery==3.1.17
tweepy==3.5.0
pyenchant==1.6.6
